+++ spec/jobs/reaction/update_records_job_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
 
-RSpec.describe ReactionCascadeJob, type: :job do
+RSpec.describe Reaction::UpdateRecordsJob, type: :job do
   pending "add some examples to (or delete) #{__FILE__}"
 end
